test-makerbase/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pyserial==3.5